requirements.txt
django 
requests
djangorestframework
click